COSI-39: rename_and_move_files_to_clients_folder
Renamed and moved iamclient and s3client files to clients/iam
and clients/s3 directories. Updated related import paths in
provisioner_server_impl.go and its test file to reflect these
changes for better organization and clarity.
